+++ internal/sdkprovider/service/kafka/mirrormaker_replication_flow_test.go
@@ -182,7 +182,7 @@ resource "aiven_kafka_mirrormaker" "mm" {
   service_name = "test-acc-sr-mm-%[2]s"
 
   kafka_mirrormaker_user_config {
-    ip_filter = ["0.0.0.0/0"]
+    ip_filter = ["0.0.0.0/0", "::/0"]
 
     kafka_mirrormaker {
       refresh_groups_interval_seconds = 600